Moved App to its own component and so index.js will only be responsible for rendering to DOM
