+++ client/src/components/App.jsx
@@ -1,0 +1,19 @@
+import React from 'react';
+import ReactDOM from 'react-dom/client';
+import App from './App.jsx';
+
+
+
+
+const App = () => {
+  return (
+    <>
+    {/* <h1 className="title">Shopping</h1>
+    <Product /> */}
+    {/* <QandA />
+    <Related /> */}
+    <Reviews />
+
+    </>
+  )
+}
